chore: Aggregate tests deps into main package.json
 * Regroup all npm-based dependencies into the package.json  of the S3 project
in order to facilitate testing in our proprietary end-to-end environment. This
will allow our dependency patcher to also easily patch the tests dependencies,
instead of forcing it to dig in order to find dependencies that require
patching for testing purposes (ie: thinking of testing alternate branches of
scality-based repositories that could be used as dependency for the tests).

 * Update the tester.js script not to run independant npm installs anymore

 * Use the latest version for regrouped dependencies