tests: Rework functional testing for improved test visibility
* Harmonize how the functional tests are launched:
  - Use local package.json to tell how to launch tests
  - Rework tester.js to use the package.json instead of the obsolete master.json files
  - Remove obsolete master.json files

* Provide convenience npm scripts for S3 to launch individual functional test suites

* Attempt to reorganize circle.yml in order to improve the visibility of each
test suite. Hopefully, thanks to this, the CircleCI UI will help us find more
quickly whichever test failed. This has of course the downside of (probably)
increasing testing time, but the visibility is a clear upside.

* Add server output redirection for individual functional tests, so as to clear
out the tests' outputs

* Add a wait script to wait for S3 to be ready before launching tests, failing
when the timeout is reached to ensure we do not wait forever.